Remove function invocation from .styles, move styles to parent
