+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/actions-time.tsx
@@ -1,5 +1,5 @@
 import VisuallyHiddenText from '../../../../components/VisuallyHiddenText';
-import styles from './index.styles';
+import styles from '../index.styles';
 import Card from './card';
 import type { PlayerActions } from '../types';
 
@@ -29,7 +29,7 @@ const ActionsTime = ({ player }: ActionsTimeProps) => {
       {times.map((time, index) => (
         // eslint-disable-next-line react/no-array-index-key
         <span aria-hidden key={index}>
-          <span css={styles.textBlock()}>
+          <span css={styles.textBlock}>
             {index === 0 && '('}
             {player.actionType === 'card' && <Card player={player} />}
             {time}
