+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/card.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/aria-role */
 import { RedCardSVG, SecondYellowCardSVG } from '#app/components/icons';
-import styles from './index.styles';
+import styles from '../index.styles';
 import type { PlayerActions } from '../types';
 
 interface CardProps {
@@ -8,14 +8,14 @@ interface CardProps {
 }
 
 const Card = ({ player }: CardProps) => (
-  <div css={styles.cardContainer()} role="text">
+  <div css={styles.cardContainer} role="text">
     {player.actions[0].type === 'Red Card' ? (
-      <span css={styles.redCard()} aria-hidden data-testid="red-card-img">
+      <span css={styles.redCard} aria-hidden data-testid="red-card-img">
         <RedCardSVG />
       </span>
     ) : (
       <span
-        css={styles.yellowCard()}
+        css={styles.yellowCard}
         aria-hidden
         data-testid="second-yellow-card-img"
       >
