+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/centre.tsx
@@ -6,7 +6,7 @@ import {
 
 import Time from './fixture-time';
 import Score from './score';
-import styles from './index.styles';
+import styles from '../index.styles';
 import type { HeadToHeadV2Data, EventStatusType } from '../types';
 
 export const shouldShowScores = (statusGroup: EventStatusType) =>
