+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/footer.tsx
@@ -1,4 +1,4 @@
-import styles from './index.styles';
+import styles from '../index.styles';
 
 interface FooterProps {
   venue: string;
@@ -10,18 +10,18 @@ const Footer = ({ venue, attendanceValue, attendanceInfo }: FooterProps) => {
   const formattedAttendanceValue = attendanceValue?.toLocaleString();
 
   return (
-    <div css={styles.footer()}>
-      <hr css={styles.horizontalRule()} aria-hidden />
-      <div css={styles.footerTextWrapper()}>
-        <div css={styles.venue()}>
-          <span css={styles.venueLabel()}>Venue:</span>
+    <div css={styles.footer}>
+      <hr css={styles.horizontalRule} aria-hidden />
+      <div css={styles.footerTextWrapper}>
+        <div css={styles.venue}>
+          <span css={styles.venueLabel}>Venue:</span>
           {attendanceInfo ? `${venue} (${attendanceInfo})` : venue}
         </div>
       </div>
-      <div css={styles.footerTextWrapper()}>
+      <div css={styles.footerTextWrapper}>
         {attendanceValue && (
-          <div css={styles.attendanceValue()}>
-            <span css={styles.attendanceLabel()}>Attendance:</span>
+          <div css={styles.attendanceValue}>
+            <span css={styles.attendanceLabel}>Attendance:</span>
             {formattedAttendanceValue}
           </div>
         )}
