+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/key-events.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import VisuallyHiddenText from '#app/components/VisuallyHiddenText';
 import Action from './action';
-import styles from './index.styles';
+import styles from '../index.styles';
 import type { PlayerActions } from '../types';
 
 interface KeyEventsProps {
@@ -19,11 +19,11 @@ export const KeyEvents = ({
 }: KeyEventsProps) => (
   <>
     <VisuallyHiddenText as="h3">Key Events</VisuallyHiddenText>
-    <div css={styles.keyEventsHome()}>
+    <div css={styles.keyEventsHome}>
       <VisuallyHiddenText as="h4">{homeName}</VisuallyHiddenText>
       <Action contestantActions={homeKeyEvents} alignment="home" />
     </div>
-    <div css={styles.keyEventsAway()}>
+    <div css={styles.keyEventsAway}>
       <VisuallyHiddenText as="h4">{awayName}</VisuallyHiddenText>
       <Action contestantActions={awayKeyEvents} alignment="away" />
     </div>
