+++ src/app/components-webcore/SportDataHeader/head-to-head-v2/components/score-details.tsx
@@ -1,5 +1,5 @@
 import VisuallyHiddenText from '../../../../components/VisuallyHiddenText';
-import styles from './index.styles';
+import styles from '../index.styles';
 import type { RunningScores } from '../types';
 
 interface ScoreDetailsProps {
@@ -30,24 +30,24 @@ const ScoreDetails = ({
   }
 
   return (
-    <div css={styles.scoreDetailsWrapper()}>
+    <div css={styles.scoreDetailsWrapper}>
       {shouldDisplayFT && (
         <>
           <VisuallyHiddenText>{`Full Time ${homeName} ${homeRunningScores?.fulltime} , ${awayName} ${awayRunningScores?.fulltime}`}</VisuallyHiddenText>
 
           <div
-            css={styles.scoreDetailsScore()}
+            css={styles.scoreDetailsScore}
             aria-hidden="true"
           >{`FT ${homeRunningScores?.fulltime}-${awayRunningScores?.fulltime}`}</div>
 
-          <span css={styles.scoreDetailsComma()}>,</span>
+          <span css={styles.scoreDetailsComma}>,</span>
         </>
       )}
       {shouldDisplayHT && (
         <>
           <VisuallyHiddenText>{`Half Time ${homeName} ${homeRunningScores?.halftime} , ${awayName} ${awayRunningScores?.halftime}`}</VisuallyHiddenText>{' '}
           <div
-            css={styles.scoreDetailsScore()}
+            css={styles.scoreDetailsScore}
             aria-hidden="true"
           >{`HT ${homeRunningScores?.halftime}-${awayRunningScores?.halftime}`}</div>
         </>
